Move DocumentExplorer components to root-content/explorer/document-explorer
