+++ src/app/root-content/explorer/document-explorer/blank-nodes/blank-node.component.html
