+++ src/app/root-content/explorer/document-explorer/blank-nodes/blank-nodes.component.scss
