+++ src/app/root-content/explorer/document-explorer/document-deleter/document-deleter.component.scss
