+++ src/app/root-content/explorer/document-explorer/document-explorer.component.scss
