+++ src/app/root-content/explorer/document-explorer/document-resource/document-resource.component.html
