+++ src/app/root-content/explorer/document-explorer/document-tree-view/document-tree-view.component.scss
