+++ src/app/root-content/explorer/document-explorer/document-viewer/document-viewer.component.html
