+++ src/app/root-content/explorer/document-explorer/documents-resolver.service.ts
