+++ src/app/root-content/explorer/document-explorer/lists/list.component.scss
