+++ src/app/root-content/explorer/document-explorer/lists/lists.component.html
