+++ src/app/root-content/explorer/document-explorer/lists/lists.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, Output, EventEmitter, OnInit } from "@angular/core";
 
 import * as RDFNode from "carbonldp/RDF/Node";
 
-import { Modes } from "./../property/property.component";
+import { Modes } from "../property/property.component";
 import { List, ListRow } from "./list.component";
 
 import "semantic-ui/semantic";
