+++ src/app/root-content/explorer/document-explorer/literals/literal.component.scss
