+++ src/app/root-content/explorer/document-explorer/literals/literals.component.scss
