+++ src/app/root-content/explorer/document-explorer/named-fragments/named-fragments.component.scss
