+++ src/app/root-content/explorer/document-explorer/pointers/pointer.component.html
