+++ src/app/root-content/explorer/document-explorer/pointers/pointers.component.html
