+++ src/app/root-content/explorer/document-explorer/pointers/pointers.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, Output, EventEmitter, OnInit } from "@angular/core";
 
 import * as RDFNode from "carbonldp/RDF/Node";
 
-import { Modes } from "./../property/property.component";
+import { Modes } from "../property/property.component";
 import { Pointer, PointerRow } from "./pointer.component";
 
 import "semantic-ui/semantic";
