+++ src/app/root-content/explorer/document-explorer/property/property.component.html
